refactor(#358): MessageList 컴포넌트 및 훅 분리 리팩토링 및 시스템 메세지 오류 수정
